safestringlib: Fix bad pointer comparisons in strcasestr_s tests
The return values from strcasestr and strcasestr_s were being cast
to int for comparison. This is bogus because pointers are only
guaranteed to fit in longs. I suspect that this was done due to compiler
warnings. Those warnings were because strcasestr was compiled
without a prototype because it is _GNU_SOURCE only, and that symbol
was not defined. So without the prototype, the return value of
strcasestr was an integer and not a pointer. I assume that the
warnings were not enabled to reveal the missing prototype.

Fix both problems.

Signed-off-by: Mark Rustad <MRustad@gmail.com>>